@@ -1,37 +1,73 @@
 {
     "name": "parse5-build-scripts",
     "private": true,
+    "type": "module",
+    "workspaces": [
+        "packages/*",
+        "bench",
+        "test"
+    ],
     "devDependencies": {
-        "dedent": "^0.7.0",
-        "eslint": "^4.19.1",
-        "eslint-config-prettier": "^2.9.0",
-        "eslint-plugin-prettier": "^2.6.0",
-        "husky": "^0.14.3",
-        "lerna": "^3.1.0",
-        "lint-staged": "^7.2.0",
-        "mocha": "^5.1.1",
-        "prettier": "^1.12.0",
-        "r2": "^2.0.1"
+        "@types/jest": "^27.4.1",
+        "@typescript-eslint/eslint-plugin": "^5.20.0",
+        "@typescript-eslint/parser": "^5.20.0",
+        "eslint": "^8.13.0",
+        "eslint-config-prettier": "^8.5.0",
+        "eslint-plugin-unicorn": "^42.0.0",
+        "husky": "^7.0.4",
+        "jest": "^27.5.1",
+        "lint-staged": "^12.3.8",
+        "outdent": "^0.8.0",
+        "prettier": "^2.6.2",
+        "ts-jest": "^27.1.4",
+        "ts-node": "^10.7.0",
+        "typedoc": "^0.22.15",
+        "typescript": "^4.6.3"
     },
     "scripts": {
-        "bootstrap": "lerna bootstrap",
-        "lint": "eslint .",
-        "unit-tests": "mocha --ui exports --reporter progress --timeout 20000 packages/**/test/*.test.js",
+        "build": "npm run build:esm && npm run build:cjs --workspaces --if-present",
+        "build:esm": "tsc --build packages/* test",
+        "build:docs": "node --max-old-space-size=8192 node_modules/.bin/typedoc .",
+        "prettier": "prettier '**/*.{js,ts,md,json,yml}' --loglevel warn",
+        "format": "npm run format:es && npm run format:prettier",
+        "format:es": "npm run lint:es -- --fix",
+        "format:prettier": "npm run prettier -- --write",
+        "lint": "npm run lint:es && npm run lint:prettier",
+        "lint:es": "eslint .",
+        "lint:prettier": "npm run prettier -- --check",
+        "unit-tests": "NODE_OPTIONS=--experimental-vm-modules jest",
         "test": "npm run lint && npm run unit-tests",
-        "generate-feedback-tests": "node scripts/generate-parser-feedback-test test/data/html5lib-tests/tree-construction/*.dat test/data/tree-construction-regression/*.dat",
-        "generate-named-entity-data": "node scripts/generate-named-entity-data",
-        "bench-perf": "cd bench && npm install && node perf/index.js",
-        "bench-memory-named-entities": "cd bench && npm install && node memory/named-entity-data.js",
-        "bench-memory-sax": "cd bench && npm install && node memory/sax-parser.js",
-        "postinstall": "npm run bootstrap",
+        "generate-feedback-tests": "node --loader ts-node/esm scripts/generate-parser-feedback-test/index.ts test/data/html5lib-tests/tree-construction/*.dat",
+        "bench-perf": "npm run build && node bench/perf/index.js",
+        "bench-memory-sax": "npm run build && node bench/memory/sax-parser.js",
         "preversion": "npm test",
-        "precommit": "lint-staged",
-        "publish": "lerna publish"
+        "pre-commit": "lint-staged",
+        "publish": "npm publish --workspaces",
+        "prepare": "husky install",
+        "prepublish": "npm run build"
     },
     "lint-staged": {
-        "*.{js,json}": [
+        "*.{js,ts}": [
             "prettier --write",
-            "git add"
+            "eslint --fix"
+        ],
+        "*.{md,json,yml}": [
+            "prettier --write"
         ]
+    },
+    "jest": {
+        "preset": "ts-jest/presets/default-esm",
+        "testEnvironment": "node",
+        "coverageProvider": "v8",
+        "globals": {
+            "ts-jest": {
+                "useESM": true
+            }
+        },
+        "moduleNameMapper": {
+            "^(parse5[^/]*)/dist/(.*?)(?:\\.js)?$": "<rootDir>/packages/$1/lib/$2",
+            "^(parse5[^/]*)$": "<rootDir>/packages/$1/lib/index.ts",
+            "^(.*)\\.js$": "$1"
+        }
     }
 }
