@@ -1,43 +1,36 @@
-'use strict';
-
-const fs = require('fs');
-const format = require('human-format');
-const promisifyEvent = require('promisify-event');
-const memwatch = require('node-memwatch');
-const SAXParser = require('../../packages/parse5-sax-parser/lib');
+import { readFile } from 'node:fs/promises';
+import format from 'human-format';
+import memwatch from '@airbnb/node-memwatch';
+import { SAXParser } from '../../packages/parse5-sax-parser/dist/index.js';
+import { finished } from 'parse5-test-utils/dist/common.js';
 
 main();
 
 async function main() {
-    let parsedDataSize = 0;
-    let maxMemUsage = 0;
-    let startDate = null;
-    let endDate = null;
     const heapDiffMeasurement = new memwatch.HeapDiff();
-    let heapDiff = null;
 
-    memwatch.on('stats', stats => {
-        maxMemUsage = Math.max(maxMemUsage, stats['current_base']);
+    let maxMemUsage = 0;
+
+    memwatch.on('stats', (stats) => {
+        maxMemUsage = Math.max(maxMemUsage, stats.used_heap_size);
     });
 
-    startDate = new Date();
+    const statsPromise = new Promise((resolve) => memwatch.once('stats', resolve));
 
-    const parserPromise = parse().then(dataSize => {
-        parsedDataSize = dataSize;
-        endDate = new Date();
-        heapDiff = heapDiffMeasurement.end();
-    });
+    const startDate = new Date();
 
-    await Promise.all([
-        parserPromise,
-        promisifyEvent(memwatch, 'stats') // NOTE: we need at least one `stats` result
-    ]);
+    const parsedDataSize = await parse();
+    const endDate = new Date();
+    const heapDiff = heapDiffMeasurement.end();
+
+    // NOTE: we need at least one `stats` result to get maxMemUsage
+    await statsPromise;
 
     printResults(parsedDataSize, startDate, endDate, heapDiff, maxMemUsage);
 }
 
 async function parse() {
-    const data = fs.readFileSync('../test/data/huge-page/huge-page.html', 'utf8');
+    const data = await readFile(new URL('../../test/data/huge-page/huge-page.html', import.meta.url), 'utf8');
     let parsedDataSize = 0;
     const stream = new SAXParser();
 
@@ -48,7 +41,7 @@ async function parse() {
 
     stream.end();
 
-    await promisifyEvent(stream, 'finish');
+    await finished(stream);
 
     return parsedDataSize;
 }
@@ -57,16 +50,16 @@ function getDuration(startDate, endDate) {
     const scale = new format.Scale({
         seconds: 1,
         minutes: 60,
-        hours: 3600
+        hours: 3600,
     });
 
-    return format((endDate - startDate) / 1000, { scale: scale });
+    return format((endDate - startDate) / 1000, { scale });
 }
 
 function printResults(parsedDataSize, startDate, endDate, heapDiff, maxMemUsage) {
     console.log('Input data size:', format(parsedDataSize, { unit: 'B' }));
-    console.log('Duration: ', getDuration(startDate, endDate));
-    console.log('Memory before: ', heapDiff.before.size);
-    console.log('Memory after: ', heapDiff.after.size);
-    console.log('Memory max: ', format(maxMemUsage, { unit: 'B' }));
+    console.log('Duration:', getDuration(startDate, endDate));
+    console.log('Memory before:', heapDiff.before.size);
+    console.log('Memory after:', heapDiff.after.size);
+    console.log('Memory max:', format(maxMemUsage, { unit: 'B' }));
 }
